Javascript Enhancements

+++ zcommerce/products/forms.py
@@ -1,6 +1,6 @@
 #customer forms.py
 from flask_wtf import FlaskForm
-from wtforms import StringField, SubmitField, IntegerField, FloatField
+from wtforms import StringField, SubmitField, IntegerField, FloatField, DecimalField
 from wtforms.validators import DataRequired, Email, NumberRange
 from wtforms import ValidationError
  
@@ -9,6 +9,6 @@ from wtforms import ValidationError
 class ProductEntry(FlaskForm):
 	product_name = StringField('Product Name', validators=[DataRequired()])
 	product_description = StringField('Description', validators=[DataRequired()])
-	product_price = FloatField('Price', validators=[DataRequired()])
+	product_price = DecimalField('Price', validators=[DataRequired()])
 	product_quantity = IntegerField('Quantity in Stock', validators=[DataRequired()])
 	submit = SubmitField('Enter')
